Split SettingsComponents' multiple components into their own file
